backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
cryptography
google-auth
google-auth-oauthlib
google-api-python-client
python-multipart
python-dotenv
pytest
httpx
pytz
caldav
requests
aiogram
